Revert "Overview page"
