--- uWeather/app/src/debug/res/mipmap-anydpi-v26/smhi_logo.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
-    <background android:drawable="@color/smhi_logo_background"/>
-    <foreground android:drawable="@mipmap/smhi_logo_foreground"/>
-</adaptive-icon>
