--- uWeather/app/src/debug/res/values/smhi_logo_background.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="smhi_logo_background">#FFFFFF</color>
-</resources>
