--- uWeather/app/src/debug/res/values/yr_logo_background.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <color name="yr_logo_background">#FFFFFF</color>
-</resources>
