+++ uWeather/app/src/main/java/dit257/mandalore/uweather/LoadingActivity.kt
@@ -6,7 +6,6 @@ import android.os.Bundle
 import android.os.Handler
 import androidx.appcompat.app.AppCompatActivity
 import dit257.mandalore.uweather.api.WeatherService
-import dit257.mandalore.uweather.manager.PreferencesManager
 import java.io.File
 
 
@@ -26,6 +25,6 @@ class LoadingActivity : AppCompatActivity() {
         // Create 10mb response cache
         HttpResponseCache.install(File(cacheDir, "http"), 10485760)
 
-        WeatherService.updateAll(PreferencesManager.getSelectedCity(this)!!)
+        WeatherService.updateAll("Gothenburg")
     }
 }
