+++ uWeather/app/src/main/java/dit257/mandalore/uweather/api/Mock2WeatherService.kt
@@ -1,0 +1,15 @@
+package dit257.mandalore.uweather.api
+
+import java.util.concurrent.Future
+
+/*class Mock2WeatherService : WeatherService("Mock 2", "") {
+    /*override fun parseResponse(response: String) {}
+
+    override fun update(lon: Float, lat: Float): Future<*>? {
+        return null
+    }
+
+    override fun getCurrentTemperature(): Double {
+        return 12.0
+    }*/
+}*/
