+++ uWeather/app/src/main/java/dit257/mandalore/uweather/api/WeatherService.kt
@@ -1,6 +1,5 @@
 package dit257.mandalore.uweather.api
 
-import dit257.mandalore.uweather.manager.CitiesManager
 import org.json.JSONException
 import org.json.JSONObject
 import java.io.IOException
@@ -32,27 +31,8 @@ abstract class WeatherService(
          * @param city the English exonym for the city at which to get the weather.
          */
         fun updateAll(city: String) {
-            val (lon, lat) = CitiesManager.getCoordinates(city)!!
+            val (lon, lat) = cities[city]!!
             services.map { it.update(lon, lat) }.forEach { it?.get() }
-        }
-
-        /**
-         * Gets the current time, zoned for UTC.
-         *
-         * @return the current time in UTC offset.
-         */
-        fun getCurrentTime(): LocalDateTime {
-            return LocalDateTime.now(ZoneOffset.UTC)
-        }
-
-        /**
-         * Calculates the average temperature from all APIs. Null values are skipped.
-         *
-         * @param time the time to get the average temperature for.
-         * @return the average temperature across all APIs at the given time.
-         */
-        fun getAverageTemperature(time: LocalDateTime): String {
-            return "%.1f°".format(services.map { it.getTemperature(time) }.filterNotNull().average())
         }
     }
 
@@ -72,7 +52,7 @@ abstract class WeatherService(
      *
      * @param lon the longitude at which to get the weather.
      * @param lat the latitude at which to get the weather.
-     * @return the [Future] for and parsing the result, or null.
+     * @return the [Future] for and parsing the result, or null
      */
     abstract fun update(lon: String, lat: String): Future<*>?
 
@@ -82,7 +62,7 @@ abstract class WeatherService(
      * @return the temperature from the API for the current time, or null.
      */
     open fun getCurrentTemperature(): Double? {
-        return getTemperature(getCurrentTime())
+        return getTemperature(LocalDateTime.now(ZoneOffset.UTC))
     }
 
     /**
