--- uWeather/app/src/main/res/layout/fragment_overview.xml
@@ -1,166 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <TextView
-        android:id="@+id/city"
-        android:layout_width="300dp"
-        android:layout_height="70dp"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="100dp"
-        android:textAlignment="center"
-        android:text="@string/city"
-        android:textColor="@color/purple_700"
-        android:textSize="48sp"
-        android:textStyle="bold" />
-
-    <TextView
-        android:id="@+id/degrees"
-        android:layout_width="300dp"
-        android:layout_height="60dp"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="170dp"
-        android:textAlignment="center"
-        android:text="@string/degrees"
-        android:textColor="@color/purple_700"
-        android:textSize="48sp"
-        android:textStyle="bold" />
-
-    <TextView
-        android:id="@+id/weather"
-        android:layout_width="300dp"
-        android:layout_height="25dp"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="230dp"
-        android:textAlignment="center"
-        android:text="@string/weather"
-        android:textColor="@color/purple_700"
-        android:textSize="16sp" />
-
-    <TextView
-        android:id="@+id/high_to_low_degrees"
-        android:layout_width="300dp"
-        android:layout_height="25dp"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="270dp"
-        android:textAlignment="center"
-        android:text="@string/high_to_low_degrees"
-        android:textColor="@color/purple_700"
-        android:textSize="16sp" />
-
-    <ImageView
-        android:id="@+id/imageView2"
-        android:layout_width="410dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="350dp"
-        android:background="@color/purple_900"
-        app:srcCompat="@drawable/hourly_forecast" />
-
-    <TextView
-        android:id="@+id/time1"
-        android:layout_width="45dp"
-        android:layout_height="25dp"
-        android:layout_marginStart="64dp"
-        android:layout_marginTop="415dp"
-        android:text="@string/time1"
-        android:textAlignment="center"
-        android:textColor="@color/white"
-        android:textSize="16sp" />
-
-    <TextView
-        android:id="@+id/time2"
-        android:layout_width="45dp"
-        android:layout_height="25dp"
-        android:layout_marginStart="123dp"
-        android:layout_marginTop="415dp"
-
-        android:text="@string/time2"
-        android:textColor="@color/white"
-        android:textSize="16sp" />
-
-    <TextView
-        android:id="@+id/time3"
-        android:layout_width="45dp"
-        android:layout_height="25dp"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="415dp"
-        android:text="@string/time3"
-        android:textAlignment="center"
-        android:textColor="@color/white"
-        android:textSize="16sp" />
-
-    <TextView
-        android:layout_gravity="center_horizontal"
-        android:id="@+id/time4"
-        android:layout_width="45dp"
-        android:layout_height="25dp"
-        android:layout_marginStart="62dp"
-        android:layout_marginTop="415dp"
-        android:text="@string/time4"
-        android:textColor="@color/white"
-        android:textSize="16sp" />
-    <TextView
-        android:id="@+id/time5"
-        android:layout_width="45dp"
-        android:layout_height="25dp"
-        android:layout_marginStart="292dp"
-        android:layout_marginTop="415dp"
-        android:text="@string/time5"
-        android:textColor="@color/white"
-        android:textSize="16sp" />
-    <TextView
-        android:id="@+id/future1"
-        android:layout_width="45dp"
-        android:layout_height="25dp"
-        android:layout_marginStart="62dp"
-        android:layout_marginTop="475dp"
-        android:textAlignment="center"
-        android:text="@string/degrees1"
-        android:textColor="@color/white"
-        android:textSize="16sp" />
-
-    <TextView
-        android:id="@+id/future2"
-        android:layout_width="45dp"
-        android:layout_height="25dp"
-        android:layout_marginStart="119dp"
-        android:layout_marginTop="475dp"
-        android:textAlignment="center"
-        android:text="@string/degrees2"
-        android:textColor="@color/white"
-        android:textSize="16sp" />
-    <TextView
-        android:id="@+id/future3"
-        android:layout_width="45dp"
-        android:layout_height="25dp"
-        android:layout_marginStart="175dp"
-        android:layout_marginTop="475dp"
-        android:textAlignment="center"
-        android:text="@string/degrees3"
-        android:textColor="@color/white"
-        android:textSize="16sp" />
-    <TextView
-        android:id="@+id/future4"
-        android:layout_width="45dp"
-        android:layout_height="25dp"
-        android:layout_marginStart="231dp"
-        android:layout_marginTop="475dp"
-        android:textAlignment="center"
-        android:text="@string/degrees4"
-        android:textColor="@color/white"
-        android:textSize="16sp" />
-
-    <TextView
-        android:id="@+id/future5"
-        android:layout_width="45dp"
-        android:layout_height="25dp"
-        android:layout_marginStart="287dp"
-        android:layout_marginTop="475dp"
-        android:text="@string/degrees4"
-        android:textAlignment="center"
-        android:textColor="@color/white"
-        android:textSize="16sp" />
-</FrameLayout>
