+++ uWeather/app/src/main/res/values-night/themes.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
     <style name="Theme.UWeather" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
