+++ uWeather/app/src/main/res/values-w1240dp/dimens.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <resources>
     <dimen name="fab_margin">200dp</dimen>
 </resources>
