+++ uWeather/app/src/main/res/values-w600dp/dimens.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <resources>
     <dimen name="fab_margin">48dp</dimen>
     <dimen name="text_margin">48dp</dimen>
